requirements.txt
git+https://github.com/huggingface/diffusers.git
accelerate
datasets
transformers
einops
ftfy
iopath
timm
pytorchvideo
kornia